+++ src/main/java/com/github/merchantpug/apugli/access/AnimalModelAccess.java
@@ -1,0 +1,11 @@
+package com.github.merchantpug.apugli.access;
+
+import net.minecraft.entity.player.PlayerEntity;
+
+public interface AnimalModelAccess {
+    PlayerEntity apugli$getPlayerEntity();
+    void apugli$setPlayerEntity(PlayerEntity playerEntity);
+
+    boolean apugli$isHidden();
+    void apugli$setHidden(boolean value);
+}